requirements.txt
flask==3.1.0
gunicorn==23.0.0
nba_api==1.5.2
xgboost==2.1.3
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.6.1
requests==2.32.3
python-dotenv==1.0.1
